template: print escape codes directly to the tty (#52)
Instead of relying on the script's stdout to be connected to the tty,
just bypass stdout and pipe the keycodes directly to the tty.

If the scripts are run in a context where stdout has been redirected (to
a file, or to /dev/null, etc), then these escape codes are lost. If the
terminal is connected to a TTY, though, there's no reason these escape
codes can't be written there anyway.